+++ test/code_size/audio_worklet_wasm.js
@@ -176,7 +176,7 @@ var N = [], O = a => {
     });
 }, ea = () => !1, fa = 1, ha = a => {
     a = a.data;
-    let b = a._wsc;
+    var b = a._wsc;
     b && C.get(b)(...a.A);
 }, ia = a => E(a), ja = (a, b, c, e, d) => {
     let g = S[a], h = g.audioWorklet, p = () => {